# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BunnyFlow AI Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register new user
 */
export const registerBodyPasswordMin = 6;

export const registerBodyUsernameMin = 3;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  username: zod.string().min(registerBodyUsernameMin),
});

/**
 * @summary Login user
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    username: zod.string(),
    plan: zod.enum(["free", "basic", "pro", "starter", "ultra", "unlimited"]),
    credits: zod.number(),
    creditsTotal: zod.number().optional(),
    planExpiresAt: zod.coerce.date().nullish(),
    daysRemaining: zod.number().nullish(),
    createdAt: zod.coerce.date(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  username: zod.string(),
  plan: zod.enum(["free", "basic", "pro", "starter", "ultra", "unlimited"]),
  credits: zod.number(),
  creditsTotal: zod.number().optional(),
  planExpiresAt: zod.coerce.date().nullish(),
  daysRemaining: zod.number().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get user credit balance
 */
export const GetCreditsResponse = zod.object({
  total: zod.number(),
  used: zod.number(),
  remaining: zod.number(),
  plan: zod.string(),
  resetDate: zod.coerce.date().optional(),
});

/**
 * @summary Get credit transaction history
 */
export const GetCreditHistoryResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["credit", "debit"]),
  amount: zod.number(),
  description: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetCreditHistoryResponse = zod.array(GetCreditHistoryResponseItem);

/**
 * @summary Get generation statistics for dashboard
 */
export const GetGenerationStatsResponse = zod.object({
  totalGenerations: zod.number(),
  videosGenerated: zod.number(),
  imagesGenerated: zod.number(),
  creditsUsedThisMonth: zod.number(),
  recentGenerations: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.enum(["video", "image"]),
      model: zod.string(),
      prompt: zod.string(),
      status: zod.enum(["pending", "processing", "completed", "failed"]),
      resultUrl: zod.string().nullish(),
      thumbnailUrl: zod.string().nullish(),
      creditsUsed: zod.number(),
      aspectRatio: zod.string(),
      duration: zod.number().nullish(),
      createdAt: zod.coerce.date(),
      completedAt: zod.coerce.date().nullish(),
    }),
  ),
  modelBreakdown: zod.array(
    zod.object({
      model: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get user's generation history
 */
export const getGenerationsQueryTypeDefault = `all`;
export const getGenerationsQueryLimitDefault = 20;
export const getGenerationsQueryOffsetDefault = 0;

export const GetGenerationsQueryParams = zod.object({
  type: zod
    .enum(["video", "image", "all"])
    .default(getGenerationsQueryTypeDefault),
  limit: zod.coerce.number().default(getGenerationsQueryLimitDefault),
  offset: zod.coerce.number().default(getGenerationsQueryOffsetDefault),
});

export const GetGenerationsResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.enum(["video", "image"]),
      model: zod.string(),
      prompt: zod.string(),
      status: zod.enum(["pending", "processing", "completed", "failed"]),
      resultUrl: zod.string().nullish(),
      thumbnailUrl: zod.string().nullish(),
      creditsUsed: zod.number(),
      aspectRatio: zod.string(),
      duration: zod.number().nullish(),
      createdAt: zod.coerce.date(),
      completedAt: zod.coerce.date().nullish(),
    }),
  ),
  total: zod.number(),
  limit: zod.number(),
  offset: zod.number(),
});

/**
 * @summary Submit a new generation request
 */

export const createGenerationBodyAspectRatioDefault = `landscape`;
export const createGenerationBodyDurationMax = 8;

export const CreateGenerationBody = zod.object({
  type: zod.enum(["video", "image"]),
  model: zod.enum([
    "veo3",
    "veo3_fast",
    "imagen4",
    "nano_banana",
    "nano_banana_pro",
  ]),
  prompt: zod.string().min(1),
  aspectRatio: zod
    .enum(["landscape", "portrait", "square"])
    .default(createGenerationBodyAspectRatioDefault),
  duration: zod
    .number()
    .min(1)
    .max(createGenerationBodyDurationMax)
    .optional()
    .describe("Duration in seconds for video (1-8)"),
});

/**
 * @summary Get a specific generation
 */
export const GetGenerationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGenerationResponse = zod.object({
  id: zod.number(),
  type: zod.enum(["video", "image"]),
  model: zod.string(),
  prompt: zod.string(),
  status: zod.enum(["pending", "processing", "completed", "failed"]),
  resultUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  creditsUsed: zod.number(),
  aspectRatio: zod.string(),
  duration: zod.number().nullish(),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullish(),
});

/**
 * @summary Get available subscription plans
 */
export const GetPlansResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  price: zod.number(),
  credits: zod.number(),
  features: zod.array(zod.string()),
  popular: zod.boolean(),
  color: zod.string(),
});
export const GetPlansResponse = zod.array(GetPlansResponseItem);